src/requirements_movie_sf.txt
#install requirements

--prefer-binary
seaborn==0.12.2
scikit-learn==1.5.2
matplotlib==3.7.0
pandas_flavor
pingouin
numpy<2
xlrd >= 2.0.1
datetime
tqdm
setuptools >=42
dill >=0.3.9
pyarrow
fastparquet
openpyxl
#geographic plot/imdb
plotly
dash[testing] pytest
dash
folium
streamlit
#GPT related func
openai
python-dotenv
geos
geopandas
click==8.0.3
sodapy
bs4
thefuzz
cinemagoer
shapely @ https://github.com/cgohlke/geospatial-wheels/releases/download/v2025.1.20/shapely-2.0.6-cp312-cp312-win_amd64.whl
